backend/requirements.txt
flask
flask-restx
flask-cors
flask-jwt-extended
flask-sqlalchemy
python-dotenv
werkzeug
